requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.1; python_version >= '3.6'
cryptography==3.4.8
flask-restful==0.3.9
flask==2.0.1
google-api-core==2.0.1; python_version >= '3.6'
google-api-python-client==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-auth==2.1.0; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
httplib2==0.19.1
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
joblib==1.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
nltk==3.6.2
numpy==1.21.2; python_version < '3.11' and python_version >= '3.7'
oauthlib==3.1.1; python_version >= '3.6'
pandas==1.3.2
protobuf==3.18.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.1.0
pymongo==3.12.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.4
pytz==2021.1
regex==2021.8.28
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
